backend/requirements.txt
fastapi
uvicorn[standard]
requests
google-generativeai
numpy
pillow
python-multipart
aiohttp
python-dotenv
